Add tests for documentation template output
This changeset also switches to snapshot testing for checking the
output of the jinja templates.
