+++ src/pyprefab/templates/docs/source/conf.py.j2
@@ -44,11 +44,11 @@ master_doc = 'index'
 
 # project information
 now = datetime.now()
-project = '{{ project_name }}'
-copyright = f'{now.year}, {{ author }} | Last update {now.strftime("%B %d, %Y")}'
-author = '{{ author }}'
+project = '{{ project_name|e }}'
+copyright = f'{now.year}, {{ author|e }} | Last update {now.strftime("%B %d, %Y")}'
+author = '{{ author|e }}'
 
-version = metadata.version('{{ project_name }}')
+version = metadata.version('{{ project_name|e }}')
 release = version
 
 # The language for content autogenerated by Sphinx. Refer to documentation
@@ -75,7 +75,7 @@ html_static_path = ['_static']
 # further. For a list of options available for each theme, see the
 # documentation.
 html_theme_options = {
-    'description': f'{{ description }}\n{release}',
+    'description': f'{{ description|e }}\n{release}',
     # 'github_user': '',
     # 'github_repo': '',
     # 'github_banner': True,
